[CSCI-3130] Assignment 2 - Todo Tasks
Refactoring the code to utilize Todo Tasks instead of Contacts.

Added removal functionality. Task removal can be done by clicking the row
in the table and clicking the "Remove" button in the resulting dialog.

Reinstated names to better match the assignment credentials. Additionally,
added changed the "date" attribute to "startDate" and "endDate" to account for
the split date attributes. Dummy data is used in the generation of dates.